Share LLM payload repair with capability probe
The probe promises to validate structured output the way production
does, but only mirrored unwrap_json — a model emitting a bare items
array or a double-encoded payload would fail qualification while its
production response succeeds. Extract the root repair into
LlmClient::PayloadRepair and apply it in both places.
